+++ CHANGELOG.md
@@ -18,6 +18,19 @@
 ## [Unreleased]
 
 
+## [v303] - 2025-08-05
+
+- Fix for checksum validation error when `NODE_BINARY_URL` is used. ([#1447](https://github.com/heroku/heroku-buildpack-nodejs/pull/1447))
+
+## [v302] - 2025-08-01
+
+- Added Node.js 24.5.0 (linux-amd64)
+- Added Node.js 22.18.0 (linux-amd64)
+
+## [v301] - 2025-07-31
+
+- Source Node.js binaries from upstream instead of mirror. ([#1441](https://github.com/heroku/heroku-buildpack-nodejs/pull/1441))
+
 ## [v300] - 2025-07-16
 
 - Added Node.js version 24.4.1.
@@ -1110,7 +1123,10 @@ Accepts `cacheDirectories` array in package.json to override default `node_modul
 
 - Documented at https://devcenter.heroku.com/articles/nodejs-support#cache-behavior
 
-[unreleased]: https://github.com/heroku/heroku-buildpack-nodejs/compare/v300...main
+[unreleased]: https://github.com/heroku/heroku-buildpack-nodejs/compare/v303...main
+[v303]: https://github.com/heroku/heroku-buildpack-nodejs/compare/v302...v303
+[v302]: https://github.com/heroku/heroku-buildpack-nodejs/compare/v301...v302
+[v301]: https://github.com/heroku/heroku-buildpack-nodejs/compare/v300...v301
 [v300]: https://github.com/heroku/heroku-buildpack-nodejs/compare/v299...v300
 [v299]: https://github.com/heroku/heroku-buildpack-nodejs/compare/v298...v299
 [v298]: https://github.com/heroku/heroku-buildpack-nodejs/compare/v297...v298